Remove unnecessary autogen.sh file
The autogen.sh script did not do anything useful, so remove it
and replace autogen.sh by autoreconf in the README.md instructions.

There is no need to create the files AUTHORS ChangeLog NEWS README:

  * configure.ac contains the `foreign` flag in AM_INIT_AUTOMAKE,
    so non-existing file would not be an issue anyway
  * The git repo already contains all these files, so they always
    exist anyway

And autoreconf is much better at running aclocal, autoheader,
automake, autoconf etc. in the proper sequence with the proper
arguments.

In addition, the autogen.sh script called configure as the last
step, which made the instructions from the README.md effectively
run configure twice, which is not necessary either.