configure.ac: Remove empty no-op [] m4 arguments
An empty TARNAME argument to AC_INIT means PACKAGE is used
as TARNAME, which is the same as with no TARNAME argument at all.

An empty ACTION-IF-NOT-FOUND argument to PKG_CONFIG_MODULES
means the same as no ACTION-IF-NOT-FOUND argument at all means,
i.e. abort configure with an error message.